Cleanup deprecated references
Basically, this package's mappings map deprecated imports to new ones.
This key was present to ease codemod and lint rules development. Indeed,
this work has been done in the past in Ember community while migrating
to ember-cli-shims and then to current imports. It has allowed us to
create ember-data-codemod and lint rules more easily. Now that it's
done, the key may be safely be removed.