Trim down mappings set
This commit removes the mappings with the key `deprecated: false`.

These mappings were duplicates kept here to have a similar structure
from what has been done in
https://github.com/ember-cli/ember-rfc176-data

The goal was to leverage was has been done for:
- https://github.com/ember-cli/ember-modules-codemod
- https://github.com/ember-cli/eslint-plugin-ember/blob/v6.7.1/docs/rules/new-module-imports.md
- https://github.com/ember-cli/eslint-plugin-ember/blob/v6.7.1/docs/rules/no-old-shims.md

to create:
- https://github.com/ember-codemods/ember-data-codemod
- the corresponding eslint rules

I've updated the codemod and the eslint rules to consume the trimmed
down set of mappings. These duplicated mappings may be safely removed
then.